refactor(ann-index): split build/update modules
